backend/requirements.txt
fastapi>=0.135.0
uvicorn>=0.41.0
openai>=2.26.0
pyyaml>=6.0.0
tavily-python>=0.7.0
requests>=2.32.0
playwright>=1.49.0
